+++ bot/commands/allow_list.rb
@@ -1,4 +1,0 @@
-module WhoIsOnDutyTodaySlackBotModule
-  module Commands
-  end
-end