+++ bot/commands/channel_labels_list.rb
@@ -1,6 +1,8 @@
 module WhoIsOnDutyTodaySlackBotModule
   module Commands
     class ChannelLabelsList
+      DESCRIPTION = 'Will display all labels in this channel.'.freeze
+      EXAMPLE = 'Usage: `channel labels list`'.freeze
       def self.call(client:, data:, match:)
         m = SlackThread.joins(:labels).where(channel_id: data.channel).group(:label).count.keys
         client.web_client.chat_postMessage(