+++ db/migrate/20241215180033_change_hide_reason_to_boolean_in_answers2.rb
@@ -1,0 +1,5 @@
+class ChangeHideReasonToBooleanInAnswers2 < ActiveRecord::Migration[7.0]
+  def change
+    change_column :answers, :hide_reason, :boolean
+  end
+end