+++ lib/opsgenie/notify_opsgenie.rb
@@ -1,40 +1,6 @@
-require "mail"
-require "yaml"
-require "erb"
 require "net/http"
 require "uri"
 require "json"
-
-class Notify
-  def initialize
-    @options = {
-      address: "replme",
-      port: 587,
-      domain: "replme"
-    }
-    @auth_options = {
-      user_name: "replme",
-      password: "replme",
-      authentication: :login,
-      enable_starttls_auto: true
-    }
-    @options.merge!(@auth_options)
-  end
-
-  def send(data)
-    @user_list = data
-
-    mail = Mail.new
-    mail.delivery_method :smtp, @options
-
-    mail.from ""
-    mail.to ""
-    mail.subject ""
-    mail.content_type "text/html; charset=UTF-8"
-    mail.body data
-    mail.deliver
-  end
-end
 
 class NotifyOpsgenie
   def initialize
@@ -48,19 +14,19 @@ class NotifyOpsgenie
     request.content_type = "application/json"
     request["Authorization"] = "GenieKey %s" % (ENV["OPSGENIE_API_TOKEN"])
     request.body = JSON.dump({
-      "message" => "#{client_info["user"]["real_name"]} calls you in slack!",
-      "description" => (message_info["permalink"]).to_s,
-      "responders" => [
-        {
-          user["field_name"] => user["name"],
-          "type" => user["type"]
-        }
-      ],
-      "tags" => [
-        "slack_bot"
-      ],
-      "priority" => "P2"
-    })
+                               "message" => "#{client_info["user"]["real_name"]} calls you in slack!",
+                               "description" => (message_info["permalink"]).to_s,
+                               "responders" => [
+                                 {
+                                   user["field_name"] => user["name"],
+                                   "type" => user["type"]
+                                 }
+                               ],
+                               "tags" => [
+                                 "slack_bot"
+                               ],
+                               "priority" => "P2"
+                             })
 
     req_options = {
       use_ssl: uri.scheme == "https"